% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd.R
\name{rd_document}
\alias{rd_document}
\title{Translate an Rd document using a specified parser}
\usage{
rd_document(
  rd,
  parser = rd_parser(),
  autolink = FALSE,
  ofile = NULL,
  level = 1
)
}
\arguments{
\item{rd}{An Rd document obtained from \code{\link[tools:parse_Rd]{parse_Rd()}}.}

\item{parser}{An \code{RdParser} object.}

\item{autolink}{A logical value indicating whether to use \code{downlit} for
automatic linking of R code documentation.}

\item{ofile}{A string specifying the output file path or a writable
\code{\link{connection}}. If \code{NULL}, the function returns the parsed document as a
character string.}

\item{level}{An integer defining the starting level of the document title.}
}
\value{
If \code{ofile} is \code{NULL}, returns a character string of the parsed
document. Otherwise, writes the output to \code{ofile} and returns
\code{invisible(ofile)}.
}
\description{
Translate an Rd document using a specified parser
}
\seealso{
\itemize{
\item \code{\link[=read_rd_source]{read_rd_source()}}/\code{\link[=read_rd_package]{read_rd_package()}}
\item \code{\link[=rd_text]{rd_text()}}/\code{\link[=rd_file]{rd_file()}}
}
}
