% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser-.R
\name{rd_parser}
\alias{rd_parser}
\title{Initialize a \code{RdParser} object}
\usage{
rd_parser(..., parser = "markdown")
}
\arguments{
\item{...}{Additional arguments for initializing the \code{\link{RdParser}}.}

\item{parser}{An \code{R6ClassGenerator} object used to initialize the
\code{\link{RdParser}}, or a string specifying the internal parser. Currently, only the
\code{"markdown"} parser is available.}
}
\value{
A \code{RdParser} object.
}
\description{
Initialize a \code{RdParser} object
}
