% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd.R
\name{rd_text}
\alias{rd_text}
\alias{rd_file}
\title{Read and parse Rd document}
\usage{
rd_text(text, ..., fragment = TRUE)

rd_file(file, ...)
}
\arguments{
\item{text}{Rd document character. Backslashes must be double-escaped
("\\\\").}

\item{...}{
  Arguments passed on to \code{\link[=rd_document]{rd_document}}
  \describe{
    \item{\code{parser}}{An \code{RdParser} object.}
    \item{\code{autolink}}{A logical value indicating whether to use \code{downlit} for
automatic linking of R code documentation.}
    \item{\code{ofile}}{A string specifying the output file path or a writable
\code{\link{connection}}. If \code{NULL}, the function returns the parsed document as a
character string.}
    \item{\code{level}}{An integer defining the starting level of the document title.}
  }}

\item{fragment}{Logical indicating whether file represents a complete
    Rd file, or a fragment.}

\item{file}{A string of the Rd document file}
}
\description{
Read and parse Rd document
}
\section{Functions}{
\itemize{
\item \code{rd_text()}: Read and parse an Rd document from a text string

\item \code{rd_file()}: Read and parse an Rd document from a file

}}
\examples{
rd_text("a\n\%b\nc")
rd_text("a & b")
rd_text("\\\\strong{x}")
}
\seealso{
\code{\link[=rd_document]{rd_document()}}
}
