% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser-.R
\name{RdParser}
\alias{RdParser}
\title{Rd Document Parser}
\description{
This R6 class provides methods for parsing and formatting Rd documentation
files. It includes functions for applying markup, structuring content, and
inserting special formatting. The class is exported to allow developers to
extend and customize the parser with additional formatting styles.
}
\details{
For Rd document sections (methods beginning with \verb{rd_*}), the internal logic
trims leading and trailing newline characters (\code{"\\\\r?\\\\n"}). Therefore, the
input for these methods consists only of the content. These methods return a
character vector, where each element corresponds to a line. Finally, a
newline character (\code{"\\n"}) is always added at the end.

The \code{tabular}, \code{ol}, \code{ul}, and \code{dl} methods accept text in a specific format
and return a character vector, where each element corresponds to a line.

The \code{subsection} Rd markup macro accepts a single-element character vector
and can also return a character vector (since it is very similar to the
\code{rd_section} method), where each element corresponds to a line. But note for
some document, it's not possible to add text after a \code{subsection} for current
section, like markdown.

For other Rd markup macros (methods not beginning with \code{"rd_"}),
they always accept a single-element character vector and should return a
character vector, where each element corresponds to a line.
}
\references{
\itemize{
\item \url{https://developer.r-project.org/parseRd.pdf}
\item \url{https://cran.r-project.org/doc/manuals/r-devel/R-exts.html#Rd-format}
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RdParser-header}{\code{RdParser$header()}}
\item \href{#method-RdParser-codeblock}{\code{RdParser$codeblock()}}
\item \href{#method-RdParser-comment}{\code{RdParser$comment()}}
\item \href{#method-RdParser-ol}{\code{RdParser$ol()}}
\item \href{#method-RdParser-ul}{\code{RdParser$ul()}}
\item \href{#method-RdParser-dl}{\code{RdParser$dl()}}
\item \href{#method-RdParser-tabular}{\code{RdParser$tabular()}}
\item \href{#method-RdParser-href}{\code{RdParser$href()}}
\item \href{#method-RdParser-email}{\code{RdParser$email()}}
\item \href{#method-RdParser-image}{\code{RdParser$image()}}
\item \href{#method-RdParser-emph}{\code{RdParser$emph()}}
\item \href{#method-RdParser-bold}{\code{RdParser$bold()}}
\item \href{#method-RdParser-strong}{\code{RdParser$strong()}}
\item \href{#method-RdParser-dQuote}{\code{RdParser$dQuote()}}
\item \href{#method-RdParser-sQuote}{\code{RdParser$sQuote()}}
\item \href{#method-RdParser-code}{\code{RdParser$code()}}
\item \href{#method-RdParser-kbd}{\code{RdParser$kbd()}}
\item \href{#method-RdParser-samp}{\code{RdParser$samp()}}
\item \href{#method-RdParser-verb}{\code{RdParser$verb()}}
\item \href{#method-RdParser-pkg}{\code{RdParser$pkg()}}
\item \href{#method-RdParser-file}{\code{RdParser$file()}}
\item \href{#method-RdParser-var}{\code{RdParser$var()}}
\item \href{#method-RdParser-env}{\code{RdParser$env()}}
\item \href{#method-RdParser-option}{\code{RdParser$option()}}
\item \href{#method-RdParser-command}{\code{RdParser$command()}}
\item \href{#method-RdParser-dfn}{\code{RdParser$dfn()}}
\item \href{#method-RdParser-cite}{\code{RdParser$cite()}}
\item \href{#method-RdParser-acronym}{\code{RdParser$acronym()}}
\item \href{#method-RdParser-abbr}{\code{RdParser$abbr()}}
\item \href{#method-RdParser-eqn}{\code{RdParser$eqn()}}
\item \href{#method-RdParser-deqn}{\code{RdParser$deqn()}}
\item \href{#method-RdParser-R}{\code{RdParser$R()}}
\item \href{#method-RdParser-dots}{\code{RdParser$dots()}}
\item \href{#method-RdParser-ldots}{\code{RdParser$ldots()}}
\item \href{#method-RdParser-cr}{\code{RdParser$cr()}}
\item \href{#method-RdParser-tab}{\code{RdParser$tab()}}
\item \href{#method-RdParser-out}{\code{RdParser$out()}}
\item \href{#method-RdParser-showif}{\code{RdParser$showif()}}
\item \href{#method-RdParser-subsection}{\code{RdParser$subsection()}}
\item \href{#method-RdParser-sectioning}{\code{RdParser$sectioning()}}
\item \href{#method-RdParser-rd_preparse}{\code{RdParser$rd_preparse()}}
\item \href{#method-RdParser-rd_postparse}{\code{RdParser$rd_postparse()}}
\item \href{#method-RdParser-rd_arguments}{\code{RdParser$rd_arguments()}}
\item \href{#method-RdParser-rd_author}{\code{RdParser$rd_author()}}
\item \href{#method-RdParser-rd_concept}{\code{RdParser$rd_concept()}}
\item \href{#method-RdParser-rd_description}{\code{RdParser$rd_description()}}
\item \href{#method-RdParser-rd_details}{\code{RdParser$rd_details()}}
\item \href{#method-RdParser-rd_docType}{\code{RdParser$rd_docType()}}
\item \href{#method-RdParser-rd_encoding}{\code{RdParser$rd_encoding()}}
\item \href{#method-RdParser-rd_format}{\code{RdParser$rd_format()}}
\item \href{#method-RdParser-rd_keyword}{\code{RdParser$rd_keyword()}}
\item \href{#method-RdParser-rd_name}{\code{RdParser$rd_name()}}
\item \href{#method-RdParser-rd_note}{\code{RdParser$rd_note()}}
\item \href{#method-RdParser-rd_references}{\code{RdParser$rd_references()}}
\item \href{#method-RdParser-rd_section}{\code{RdParser$rd_section()}}
\item \href{#method-RdParser-rd_seealso}{\code{RdParser$rd_seealso()}}
\item \href{#method-RdParser-rd_source}{\code{RdParser$rd_source()}}
\item \href{#method-RdParser-rd_title}{\code{RdParser$rd_title()}}
\item \href{#method-RdParser-rd_value}{\code{RdParser$rd_value()}}
\item \href{#method-RdParser-rd_examples}{\code{RdParser$rd_examples()}}
\item \href{#method-RdParser-rd_usage}{\code{RdParser$rd_usage()}}
\item \href{#method-RdParser-rd_alias}{\code{RdParser$rd_alias()}}
\item \href{#method-RdParser-rd_Rdversion}{\code{RdParser$rd_Rdversion()}}
\item \href{#method-RdParser-rd_synopsis}{\code{RdParser$rd_synopsis()}}
\item \href{#method-RdParser-rd_Sexpr}{\code{RdParser$rd_Sexpr()}}
\item \href{#method-RdParser-rd_RdOpts}{\code{RdParser$rd_RdOpts()}}
\item \href{#method-RdParser-clone}{\code{RdParser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-header"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-header}{}}}
\subsection{Method \code{header()}}{
Indicate a section header
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$header(text, level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}

\item{\code{level}}{An integer representing the header level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-codeblock"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-codeblock}{}}}
\subsection{Method \code{codeblock()}}{
Add a code block
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$codeblock(text, lang = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}

\item{\code{lang}}{The language of the code block. Usually \code{R}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-comment"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-comment}{}}}
\subsection{Method \code{comment()}}{
Add a comment block
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$comment(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-ol"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-ol}{}}}
\subsection{Method \code{ol()}}{
Create an ordered list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$ol(texts)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{texts}}{A character vector of multiple text elements.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character of each line for the list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-ul"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-ul}{}}}
\subsection{Method \code{ul()}}{
Create an unordered list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$ul(texts)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{texts}}{A character vector of multiple text elements.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character of each line for the list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-dl"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-dl}{}}}
\subsection{Method \code{dl()}}{
Create a definition list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$dl(terms, descriptions)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{terms}}{A character vector of terms.}

\item{\code{descriptions}}{A character vector of term definitions.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character of each line for the list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-tabular"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-tabular}{}}}
\subsection{Method \code{tabular()}}{
Create a tabular representation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$tabular(table, align)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table}}{A matrix of the table content.}

\item{\code{align}}{An character of \code{"right"}, \code{"center"}, \code{"left"} indicate
the alignment for each column in the table.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character of each row for the table.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-href"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-href}{}}}
\subsection{Method \code{href()}}{
Create a hyperlink
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$href(text, href = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}

\item{\code{href}}{A string specifying the target URL.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-email"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-email}{}}}
\subsection{Method \code{email()}}{
Create an email link
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$email(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-image"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-image}{}}}
\subsection{Method \code{image()}}{
Insert an image
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$image(src, alt = NULL, options = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{src}}{The image source path or URL.}

\item{\code{alt}}{Optional. Alternative text for the image.}

\item{\code{options}}{Optional. Additional formatting options.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-emph"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-emph}{}}}
\subsection{Method \code{emph()}}{
Apply italic formatting
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$emph(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-bold"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-bold}{}}}
\subsection{Method \code{bold()}}{
Apply bold formatting
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$bold(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-strong"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-strong}{}}}
\subsection{Method \code{strong()}}{
Apply italic and bold formatting
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$strong(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-dQuote"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-dQuote}{}}}
\subsection{Method \code{dQuote()}}{
Apply double quotes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$dQuote(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-sQuote"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-sQuote}{}}}
\subsection{Method \code{sQuote()}}{
Apply single quotes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$sQuote(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-code"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-code}{}}}
\subsection{Method \code{code()}}{
Format a fragment of R code or the name of an R object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$code(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-kbd"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-kbd}{}}}
\subsection{Method \code{kbd()}}{
Format text as keyboard input
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$kbd(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-samp"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-samp}{}}}
\subsection{Method \code{samp()}}{
Indicate verbatim word-wrapped text
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$samp(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-verb"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-verb}{}}}
\subsection{Method \code{verb()}}{
Indicate verbatim text (no wrapping or reformatting)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$verb(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-pkg"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-pkg}{}}}
\subsection{Method \code{pkg()}}{
Indicate the name of an R package
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$pkg(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-file"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-file}{}}}
\subsection{Method \code{file()}}{
Indicate a file path
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$file(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-var"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-var}{}}}
\subsection{Method \code{var()}}{
Indicate a metasyntactic variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$var(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-env"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-env}{}}}
\subsection{Method \code{env()}}{
Indicate environment variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$env(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-option"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-option}{}}}
\subsection{Method \code{option()}}{
Indicate a command-line option
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$option(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-command"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-command}{}}}
\subsection{Method \code{command()}}{
Indicate the name of a command
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$command(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-dfn"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-dfn}{}}}
\subsection{Method \code{dfn()}}{
Indicate the introductory or defining use of a term
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$dfn(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-cite"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-cite}{}}}
\subsection{Method \code{cite()}}{
Indicate a reference without a direct cross-reference
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$cite(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-acronym"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-acronym}{}}}
\subsection{Method \code{acronym()}}{
Indicate an acronym (an abbreviation written in all
capital letters)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$acronym(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-abbr"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-abbr}{}}}
\subsection{Method \code{abbr()}}{
Indicate abbreviation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$abbr(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-eqn"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-eqn}{}}}
\subsection{Method \code{eqn()}}{
Inline equation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$eqn(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-deqn"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-deqn}{}}}
\subsection{Method \code{deqn()}}{
Display (block) equation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$deqn(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-R"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-R}{}}}
\subsection{Method \code{R()}}{
Insert the R system itself
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$R()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-dots"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-dots}{}}}
\subsection{Method \code{dots()}}{
Insert an ellipsis of function argument lists
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$dots()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-ldots"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-ldots}{}}}
\subsection{Method \code{ldots()}}{
Insert an ellipsis of ordinary text.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$ldots()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-cr"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-cr}{}}}
\subsection{Method \code{cr()}}{
Insert a line break
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$cr()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-tab"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-tab}{}}}
\subsection{Method \code{tab()}}{
Insert a tab character
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$tab()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-out"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-out}{}}}
\subsection{Method \code{out()}}{
Indicate literal text
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$out(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-showif"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-showif}{}}}
\subsection{Method \code{showif()}}{
Conditional text
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$showif(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{Input format, currently the following formats are
recognized: \code{example}, \code{html}, \code{latex} and \code{text}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A single boolean value indicates whether the text should be
displayed for current \code{if} and \code{ifelse} Rd block.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-subsection"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-subsection}{}}}
\subsection{Method \code{subsection()}}{
Format the subsection content. \code{subsection} is a markup
macro used within sections. Therefore, \code{rd_preparse} and
\code{rd_postparse} won't be applied before or after this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$subsection(text, title, level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}

\item{\code{title}}{A string of the subsection title.}

\item{\code{level}}{An integer representing the header level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-sectioning"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-sectioning}{}}}
\subsection{Method \code{sectioning()}}{
Formats the Rd document sectioning. Some sections may
contain multiple paragraphs and nested blocks. This default method
just ensures that each block follows a newline character.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$sectioning(parsed_list, docs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parsed_list}}{A list of one-element string of the parsed text.}

\item{\code{docs}}{A list of the raw Rd document text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-rd_preparse"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-rd_preparse}{}}}
\subsection{Method \code{rd_preparse()}}{
A hook of command to be run before parsing the Rd
section
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$rd_preparse(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-rd_postparse"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-rd_postparse}{}}}
\subsection{Method \code{rd_postparse()}}{
A hook of command to be run after parsing the Rd section
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$rd_postparse(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string of formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-rd_arguments"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-rd_arguments}{}}}
\subsection{Method \code{rd_arguments()}}{
Format the argument setction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$rd_arguments(text, level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}

\item{\code{level}}{An integer representing the header level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-rd_author"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-rd_author}{}}}
\subsection{Method \code{rd_author()}}{
Format the author setction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$rd_author(text, level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}

\item{\code{level}}{An integer representing the header level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-rd_concept"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-rd_concept}{}}}
\subsection{Method \code{rd_concept()}}{
Format the concept setction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$rd_concept(text, level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}

\item{\code{level}}{An integer representing the header level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-rd_description"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-rd_description}{}}}
\subsection{Method \code{rd_description()}}{
Format the description setction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$rd_description(text, level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}

\item{\code{level}}{An integer representing the header level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-rd_details"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-rd_details}{}}}
\subsection{Method \code{rd_details()}}{
Format the details setction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$rd_details(text, level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}

\item{\code{level}}{An integer representing the header level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-rd_docType"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-rd_docType}{}}}
\subsection{Method \code{rd_docType()}}{
Format the docType setction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$rd_docType(text, level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}

\item{\code{level}}{An integer representing the header level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-rd_encoding"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-rd_encoding}{}}}
\subsection{Method \code{rd_encoding()}}{
Format the encoding setction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$rd_encoding(text, level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}

\item{\code{level}}{An integer representing the header level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-rd_format"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-rd_format}{}}}
\subsection{Method \code{rd_format()}}{
Style the format setction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$rd_format(text, level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}

\item{\code{level}}{An integer representing the header level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-rd_keyword"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-rd_keyword}{}}}
\subsection{Method \code{rd_keyword()}}{
Format the keyword setction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$rd_keyword(text, level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}

\item{\code{level}}{An integer representing the header level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-rd_name"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-rd_name}{}}}
\subsection{Method \code{rd_name()}}{
Format the name setction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$rd_name(text, level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}

\item{\code{level}}{An integer representing the header level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-rd_note"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-rd_note}{}}}
\subsection{Method \code{rd_note()}}{
Format the note setction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$rd_note(text, level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}

\item{\code{level}}{An integer representing the header level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-rd_references"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-rd_references}{}}}
\subsection{Method \code{rd_references()}}{
Format the references setction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$rd_references(text, level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}

\item{\code{level}}{An integer representing the header level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-rd_section"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-rd_section}{}}}
\subsection{Method \code{rd_section()}}{
Format the section content
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$rd_section(text, title, level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}

\item{\code{title}}{A string of the section title.}

\item{\code{level}}{An integer representing the header level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-rd_seealso"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-rd_seealso}{}}}
\subsection{Method \code{rd_seealso()}}{
Format the seealso section
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$rd_seealso(text, level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}

\item{\code{level}}{An integer representing the header level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-rd_source"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-rd_source}{}}}
\subsection{Method \code{rd_source()}}{
Format the source section
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$rd_source(text, level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}

\item{\code{level}}{An integer representing the header level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-rd_title"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-rd_title}{}}}
\subsection{Method \code{rd_title()}}{
Format the title section
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$rd_title(text, level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}

\item{\code{level}}{An integer representing the header level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-rd_value"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-rd_value}{}}}
\subsection{Method \code{rd_value()}}{
Format the value section
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$rd_value(text, level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}

\item{\code{level}}{An integer representing the header level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-rd_examples"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-rd_examples}{}}}
\subsection{Method \code{rd_examples()}}{
Format the examples section
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$rd_examples(text, level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}

\item{\code{level}}{An integer representing the header level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-rd_usage"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-rd_usage}{}}}
\subsection{Method \code{rd_usage()}}{
Format the usage section
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$rd_usage(text, level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}

\item{\code{level}}{An integer representing the header level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-rd_alias"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-rd_alias}{}}}
\subsection{Method \code{rd_alias()}}{
Format the alias section
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$rd_alias(text, level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}

\item{\code{level}}{An integer representing the header level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-rd_Rdversion"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-rd_Rdversion}{}}}
\subsection{Method \code{rd_Rdversion()}}{
Format the Rdversion section
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$rd_Rdversion(text, level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}

\item{\code{level}}{An integer representing the header level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-rd_synopsis"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-rd_synopsis}{}}}
\subsection{Method \code{rd_synopsis()}}{
Format the synopsis section
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$rd_synopsis(text, level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}

\item{\code{level}}{An integer representing the header level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-rd_Sexpr"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-rd_Sexpr}{}}}
\subsection{Method \code{rd_Sexpr()}}{
Format the Sexpr section
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$rd_Sexpr(text, level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}

\item{\code{level}}{An integer representing the header level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-rd_RdOpts"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-rd_RdOpts}{}}}
\subsection{Method \code{rd_RdOpts()}}{
Format the RdOpts section
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$rd_RdOpts(text, level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A one-element character representing the input text.}

\item{\code{level}}{An integer representing the header level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The formatted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParser-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RdParser-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
