% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser-markdown.R
\name{RdParserMarkdown}
\alias{RdParserMarkdown}
\title{Rd document parser in pandoc markdown style}
\description{
A \code{\link{RdParser}} generator for pandoc markdown style.
}
\keyword{internal}
\section{Super class}{
\code{\link[rdparser:RdParser]{rdparser::RdParser}} -> \code{RdParserMarkdown}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RdParserMarkdown-header}{\code{RdParserMarkdown$header()}}
\item \href{#method-RdParserMarkdown-codeblock}{\code{RdParserMarkdown$codeblock()}}
\item \href{#method-RdParserMarkdown-comment}{\code{RdParserMarkdown$comment()}}
\item \href{#method-RdParserMarkdown-ol}{\code{RdParserMarkdown$ol()}}
\item \href{#method-RdParserMarkdown-ul}{\code{RdParserMarkdown$ul()}}
\item \href{#method-RdParserMarkdown-dl}{\code{RdParserMarkdown$dl()}}
\item \href{#method-RdParserMarkdown-tabular}{\code{RdParserMarkdown$tabular()}}
\item \href{#method-RdParserMarkdown-href}{\code{RdParserMarkdown$href()}}
\item \href{#method-RdParserMarkdown-image}{\code{RdParserMarkdown$image()}}
\item \href{#method-RdParserMarkdown-emph}{\code{RdParserMarkdown$emph()}}
\item \href{#method-RdParserMarkdown-bold}{\code{RdParserMarkdown$bold()}}
\item \href{#method-RdParserMarkdown-strong}{\code{RdParserMarkdown$strong()}}
\item \href{#method-RdParserMarkdown-code}{\code{RdParserMarkdown$code()}}
\item \href{#method-RdParserMarkdown-eqn}{\code{RdParserMarkdown$eqn()}}
\item \href{#method-RdParserMarkdown-deqn}{\code{RdParserMarkdown$deqn()}}
\item \href{#method-RdParserMarkdown-cr}{\code{RdParserMarkdown$cr()}}
\item \href{#method-RdParserMarkdown-showif}{\code{RdParserMarkdown$showif()}}
\item \href{#method-RdParserMarkdown-clone}{\code{RdParserMarkdown$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="R"><a href='../../rdparser/html/RdParser.html#method-RdParser-R'><code>rdparser::RdParser$R()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="abbr"><a href='../../rdparser/html/RdParser.html#method-RdParser-abbr'><code>rdparser::RdParser$abbr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="acronym"><a href='../../rdparser/html/RdParser.html#method-RdParser-acronym'><code>rdparser::RdParser$acronym()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="cite"><a href='../../rdparser/html/RdParser.html#method-RdParser-cite'><code>rdparser::RdParser$cite()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="command"><a href='../../rdparser/html/RdParser.html#method-RdParser-command'><code>rdparser::RdParser$command()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="dQuote"><a href='../../rdparser/html/RdParser.html#method-RdParser-dQuote'><code>rdparser::RdParser$dQuote()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="dfn"><a href='../../rdparser/html/RdParser.html#method-RdParser-dfn'><code>rdparser::RdParser$dfn()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="dots"><a href='../../rdparser/html/RdParser.html#method-RdParser-dots'><code>rdparser::RdParser$dots()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="email"><a href='../../rdparser/html/RdParser.html#method-RdParser-email'><code>rdparser::RdParser$email()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="env"><a href='../../rdparser/html/RdParser.html#method-RdParser-env'><code>rdparser::RdParser$env()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="file"><a href='../../rdparser/html/RdParser.html#method-RdParser-file'><code>rdparser::RdParser$file()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="kbd"><a href='../../rdparser/html/RdParser.html#method-RdParser-kbd'><code>rdparser::RdParser$kbd()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="ldots"><a href='../../rdparser/html/RdParser.html#method-RdParser-ldots'><code>rdparser::RdParser$ldots()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="option"><a href='../../rdparser/html/RdParser.html#method-RdParser-option'><code>rdparser::RdParser$option()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="out"><a href='../../rdparser/html/RdParser.html#method-RdParser-out'><code>rdparser::RdParser$out()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="pkg"><a href='../../rdparser/html/RdParser.html#method-RdParser-pkg'><code>rdparser::RdParser$pkg()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="rd_RdOpts"><a href='../../rdparser/html/RdParser.html#method-RdParser-rd_RdOpts'><code>rdparser::RdParser$rd_RdOpts()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="rd_Rdversion"><a href='../../rdparser/html/RdParser.html#method-RdParser-rd_Rdversion'><code>rdparser::RdParser$rd_Rdversion()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="rd_Sexpr"><a href='../../rdparser/html/RdParser.html#method-RdParser-rd_Sexpr'><code>rdparser::RdParser$rd_Sexpr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="rd_alias"><a href='../../rdparser/html/RdParser.html#method-RdParser-rd_alias'><code>rdparser::RdParser$rd_alias()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="rd_arguments"><a href='../../rdparser/html/RdParser.html#method-RdParser-rd_arguments'><code>rdparser::RdParser$rd_arguments()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="rd_author"><a href='../../rdparser/html/RdParser.html#method-RdParser-rd_author'><code>rdparser::RdParser$rd_author()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="rd_concept"><a href='../../rdparser/html/RdParser.html#method-RdParser-rd_concept'><code>rdparser::RdParser$rd_concept()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="rd_description"><a href='../../rdparser/html/RdParser.html#method-RdParser-rd_description'><code>rdparser::RdParser$rd_description()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="rd_details"><a href='../../rdparser/html/RdParser.html#method-RdParser-rd_details'><code>rdparser::RdParser$rd_details()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="rd_docType"><a href='../../rdparser/html/RdParser.html#method-RdParser-rd_docType'><code>rdparser::RdParser$rd_docType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="rd_encoding"><a href='../../rdparser/html/RdParser.html#method-RdParser-rd_encoding'><code>rdparser::RdParser$rd_encoding()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="rd_examples"><a href='../../rdparser/html/RdParser.html#method-RdParser-rd_examples'><code>rdparser::RdParser$rd_examples()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="rd_format"><a href='../../rdparser/html/RdParser.html#method-RdParser-rd_format'><code>rdparser::RdParser$rd_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="rd_keyword"><a href='../../rdparser/html/RdParser.html#method-RdParser-rd_keyword'><code>rdparser::RdParser$rd_keyword()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="rd_name"><a href='../../rdparser/html/RdParser.html#method-RdParser-rd_name'><code>rdparser::RdParser$rd_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="rd_note"><a href='../../rdparser/html/RdParser.html#method-RdParser-rd_note'><code>rdparser::RdParser$rd_note()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="rd_postparse"><a href='../../rdparser/html/RdParser.html#method-RdParser-rd_postparse'><code>rdparser::RdParser$rd_postparse()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="rd_preparse"><a href='../../rdparser/html/RdParser.html#method-RdParser-rd_preparse'><code>rdparser::RdParser$rd_preparse()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="rd_references"><a href='../../rdparser/html/RdParser.html#method-RdParser-rd_references'><code>rdparser::RdParser$rd_references()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="rd_section"><a href='../../rdparser/html/RdParser.html#method-RdParser-rd_section'><code>rdparser::RdParser$rd_section()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="rd_seealso"><a href='../../rdparser/html/RdParser.html#method-RdParser-rd_seealso'><code>rdparser::RdParser$rd_seealso()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="rd_source"><a href='../../rdparser/html/RdParser.html#method-RdParser-rd_source'><code>rdparser::RdParser$rd_source()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="rd_synopsis"><a href='../../rdparser/html/RdParser.html#method-RdParser-rd_synopsis'><code>rdparser::RdParser$rd_synopsis()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="rd_title"><a href='../../rdparser/html/RdParser.html#method-RdParser-rd_title'><code>rdparser::RdParser$rd_title()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="rd_usage"><a href='../../rdparser/html/RdParser.html#method-RdParser-rd_usage'><code>rdparser::RdParser$rd_usage()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="rd_value"><a href='../../rdparser/html/RdParser.html#method-RdParser-rd_value'><code>rdparser::RdParser$rd_value()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="sQuote"><a href='../../rdparser/html/RdParser.html#method-RdParser-sQuote'><code>rdparser::RdParser$sQuote()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="samp"><a href='../../rdparser/html/RdParser.html#method-RdParser-samp'><code>rdparser::RdParser$samp()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="sectioning"><a href='../../rdparser/html/RdParser.html#method-RdParser-sectioning'><code>rdparser::RdParser$sectioning()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="subsection"><a href='../../rdparser/html/RdParser.html#method-RdParser-subsection'><code>rdparser::RdParser$subsection()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="tab"><a href='../../rdparser/html/RdParser.html#method-RdParser-tab'><code>rdparser::RdParser$tab()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="var"><a href='../../rdparser/html/RdParser.html#method-RdParser-var'><code>rdparser::RdParser$var()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rdparser" data-topic="RdParser" data-id="verb"><a href='../../rdparser/html/RdParser.html#method-RdParser-verb'><code>rdparser::RdParser$verb()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParserMarkdown-header"></a>}}
\if{latex}{\out{\hypertarget{method-RdParserMarkdown-header}{}}}
\subsection{Method \code{header()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParserMarkdown$header(text, level)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParserMarkdown-codeblock"></a>}}
\if{latex}{\out{\hypertarget{method-RdParserMarkdown-codeblock}{}}}
\subsection{Method \code{codeblock()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParserMarkdown$codeblock(text, lang = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParserMarkdown-comment"></a>}}
\if{latex}{\out{\hypertarget{method-RdParserMarkdown-comment}{}}}
\subsection{Method \code{comment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParserMarkdown$comment(text)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParserMarkdown-ol"></a>}}
\if{latex}{\out{\hypertarget{method-RdParserMarkdown-ol}{}}}
\subsection{Method \code{ol()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParserMarkdown$ol(texts)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParserMarkdown-ul"></a>}}
\if{latex}{\out{\hypertarget{method-RdParserMarkdown-ul}{}}}
\subsection{Method \code{ul()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParserMarkdown$ul(texts)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParserMarkdown-dl"></a>}}
\if{latex}{\out{\hypertarget{method-RdParserMarkdown-dl}{}}}
\subsection{Method \code{dl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParserMarkdown$dl(terms, descriptions)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParserMarkdown-tabular"></a>}}
\if{latex}{\out{\hypertarget{method-RdParserMarkdown-tabular}{}}}
\subsection{Method \code{tabular()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParserMarkdown$tabular(table, align)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParserMarkdown-href"></a>}}
\if{latex}{\out{\hypertarget{method-RdParserMarkdown-href}{}}}
\subsection{Method \code{href()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParserMarkdown$href(text, href = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParserMarkdown-image"></a>}}
\if{latex}{\out{\hypertarget{method-RdParserMarkdown-image}{}}}
\subsection{Method \code{image()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParserMarkdown$image(src, alt = NULL, options = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParserMarkdown-emph"></a>}}
\if{latex}{\out{\hypertarget{method-RdParserMarkdown-emph}{}}}
\subsection{Method \code{emph()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParserMarkdown$emph(text)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParserMarkdown-bold"></a>}}
\if{latex}{\out{\hypertarget{method-RdParserMarkdown-bold}{}}}
\subsection{Method \code{bold()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParserMarkdown$bold(text)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParserMarkdown-strong"></a>}}
\if{latex}{\out{\hypertarget{method-RdParserMarkdown-strong}{}}}
\subsection{Method \code{strong()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParserMarkdown$strong(text)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParserMarkdown-code"></a>}}
\if{latex}{\out{\hypertarget{method-RdParserMarkdown-code}{}}}
\subsection{Method \code{code()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParserMarkdown$code(text)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParserMarkdown-eqn"></a>}}
\if{latex}{\out{\hypertarget{method-RdParserMarkdown-eqn}{}}}
\subsection{Method \code{eqn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParserMarkdown$eqn(text)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParserMarkdown-deqn"></a>}}
\if{latex}{\out{\hypertarget{method-RdParserMarkdown-deqn}{}}}
\subsection{Method \code{deqn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParserMarkdown$deqn(text)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParserMarkdown-cr"></a>}}
\if{latex}{\out{\hypertarget{method-RdParserMarkdown-cr}{}}}
\subsection{Method \code{cr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParserMarkdown$cr()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParserMarkdown-showif"></a>}}
\if{latex}{\out{\hypertarget{method-RdParserMarkdown-showif}{}}}
\subsection{Method \code{showif()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParserMarkdown$showif(text)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RdParserMarkdown-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RdParserMarkdown-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdParserMarkdown$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
