% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd.R
\name{read_rd_source}
\alias{read_rd_source}
\alias{read_rd_package}
\title{Read Rd document}
\usage{
read_rd_source(source)

read_rd_package(package, lib = NULL)
}
\arguments{
\item{source}{A string of the directory of a R source package.}

\item{package}{A string of the installed package name.}

\item{lib}{A character vector of directory names of R libraries.  The default
value of \code{NULL} corresponds to all libraries currently known. The specified
library trees are used to search for package.}
}
\value{
A list of \code{\link[tools:parse_Rd]{Rd}} object.
}
\description{
Reads an Rd document, which can then be translated using \code{\link[=rd_document]{rd_document()}}.
}
\section{Functions}{
\itemize{
\item \code{read_rd_source()}: Read Rd document from a R source package

\item \code{read_rd_package()}: Read Rd document from an installed package.

}}
\seealso{
\code{\link[=rd_document]{rd_document()}}
}
